Rename gestureHandlers to gestureHandlerCommon
