+++ src/components/Swipeable.tsx
@@ -22,7 +22,7 @@ import {
   PanGestureHandlerEventPayload,
   HandlerStateChangeEvent,
   TapGestureHandlerEventPayload,
-} from '../handlers/gestureHandlers';
+} from '../handlers/gestureHandlerCommon';
 import { State } from '../State';
 
 const DRAG_TOSS = 0.05;
