+++ src/handlers/gestureHandlerCommon.ts
